Adding php to insert into database

+++ web/The Worst/addDB.php
@@ -4,7 +4,7 @@
     $mainpicture = $_POST['mainPicture'];    
     $previewpicture = $_POST['previewPicture'];
     $sidepicture = $_POST['sidePicture'];
-    $price = (float)$_POST['price'];
+    $price = (double)$_POST['price'];
     $section = (int)$_POST['section'];
